Moved FastHashMap and friends
